requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
gTTS>=2.5.1
SpeechRecognition>=3.10.1
pytesseract>=0.3.10
python-dotenv>=1.0.1
googlemaps>=4.10.0
Pillow>=10.2.0
numpy>=1.26.4
pandas>=2.2.1
python-dateutil>=2.8.2
requests>=2.31.0
setuptools>=65.6.3
gunicorn>=21.2.0
pdfplumber>=0.10.3
PyMuPDF>=1.23.26